requirements.txt
django
djangorestframework
django-cors-headers
drf-yasg
celery
rabbitmq
django-environ
mysql-connector-python
